requirements.txt
openai>=1.30.0
chromadb>=0.5.0
pdfplumber>=0.11.0
streamlit>=1.32.0
python-dotenv>=1.0.0
